+++ export_readiness/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 
 from wagtail.core.models import Page
+from wagtail.core import blocks as wagtail_blocks
 
 from modelcluster.fields import ParentalManyToManyField
 
@@ -17,6 +18,7 @@ from core.constants import ARTICLE_TYPES
 from core.mixins import ServiceHomepageMixin, ServiceNameUniqueSlugMixin
 
 from . import panels, snippets
+from core import blocks, fields
 
 
 class BaseDomesticPage(ServiceNameUniqueSlugMixin, BasePage):
@@ -1340,6 +1342,46 @@ class HomePage(
     banner_label = models.CharField(max_length=50, null=True, blank=True)
     news_title = models.CharField(max_length=255)
     news_description = MarkdownField()
+
+    # hero
+    hero_image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+    hero_text = models.TextField(null=True, blank=True)
+    hero_cta_text = models.CharField(null=True, blank=True, max_length=255)
+    hero_cta_linked_page = models.ForeignKey(
+        'wagtailcore.Page',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+
+    # how DIT helps
+    how_dit_helps_title = models.TextField(null=True, blank=True)
+    how_dit_helps_columns = fields.single_struct_block_stream_field_factory(
+        field_name='columns',
+        block_class_instance=blocks.ColumnWithTitleIconTextBlock(),
+        max_num=3, null=True, blank=True
+    )
+    # questions accordions
+    questions_section_title = models.CharField(max_length=255, null=True, blank=True)
+    questions = fields.single_struct_block_stream_field_factory(
+        field_name='questions',
+        block_class_instance=blocks.DetailsSummaryBlock(),
+        max_num=5, null=True, blank=True
+    )
+    # what's new
+    what_is_new_title = models.CharField(max_length=255, null=True, blank=True)
+    what_is_new_pages = fields.single_struct_block_stream_field_factory(
+        field_name='pages',
+        block_class_instance=wagtail_blocks.PageChooserBlock(page_type='export_readiness.ArticlePage'),
+        max_num=6, null=True, blank=True
+    )
 
     @staticmethod
     def get_verbose_name():