+++ tests/great_international/test_views.py
@@ -1,10 +1,11 @@
 from directory_constants import cms
-
+import pytest
 from rest_framework.reverse import reverse
 
 from tests.great_international import factories
 
 
+@pytest.mark.django_db
 def test_international_sector_page(admin_client, international_root_page):
     sector_page = factories.InternationalSectorPageFactory.create(
         parent=international_root_page
@@ -14,6 +15,7 @@ def test_international_sector_page(admin_client, international_root_page):
     assert response.status_code == 200
 
 
+@pytest.mark.django_db
 def test_international_homepage(admin_client, root_page):
     home_page = factories.InternationalHomePageFactory.create(
         parent=root_page
@@ -27,8 +29,8 @@ def test_international_homepage(admin_client, root_page):
     assert list(featured_links.keys()) == ['image', 'heading']
 
 
+@pytest.mark.django_db
 def test_international_campaign_page(admin_client, international_root_page):
-
     campaign_page = factories.InternationalCampaignPageFactory(
         parent=international_root_page
     )
@@ -37,6 +39,7 @@ def test_international_campaign_page(admin_client, international_root_page):
     assert response.status_code == 200
 
 
+@pytest.mark.django_db
 def test_international_article_listing_page_view(admin_client, international_root_page):
     article_listing_page = factories.InternationalArticleListingPageFactory.create(  # NOQA
         parent=international_root_page,
@@ -64,6 +67,7 @@ def test_international_article_listing_page_view(admin_client, international_roo
     assert 'meta' in response.json()['child_pages'][0]
 
 
+@pytest.mark.django_db
 def test_international_topic_landing_page_view(admin_client, international_root_page):
     topic_landing_page = factories.InternationalTopicLandingPageFactory.create(
         parent=international_root_page,
@@ -73,7 +77,7 @@ def test_international_topic_landing_page_view(admin_client, international_root_
         parent=topic_landing_page,
         live=True
     )
-    article_listing_page = factories.InternationalArticleListingPageFactory.create(  # NOQA
+    article_listing_page = factories.InternationalArticleListingPageFactory.create(
         parent=topic_landing_page,
         live=True
     )
@@ -88,6 +92,7 @@ def test_international_topic_landing_page_view(admin_client, international_root_
     ) == [article_listing_page.pk, campaign_page.pk]
 
 
+@pytest.mark.django_db
 def test_international_topic_landing_page_view_sectors_alphabetical_order(
         admin_client, international_root_page):
     landing_page = factories.InternationalTopicLandingPageFactory.create(
@@ -111,104 +116,7 @@ def test_international_topic_landing_page_view_sectors_alphabetical_order(
     ]
 
 
-def test_article_listingpage_with_localised_content(admin_client, international_root_page):
-    article_listing_page = factories.InternationalArticleListingPageFactory.create(  # NOQA
-        parent=international_root_page,
-        live=True,
-        slug='setup-uk'
-    )
-    factories.InternationalArticlePageFactory.create(
-        parent=article_listing_page,
-        live=True
-    )
-    region_page = factories.InternationalRegionPageFactory(
-        parent=international_root_page,
-        live=True,
-        slug='germany'
-    )
-    regional_folder_page = factories.InternationalLocalisedFolderPageFactory(
-        parent=region_page,
-        live=True,
-        slug='setup-uk'
-    )
-    localised_article = factories.InternationalArticlePageFactory.create(
-        parent=regional_folder_page,
-        live=True
-    )
-
-    url = reverse(
-        'api:lookup-by-slug',
-        kwargs={'slug': article_listing_page.slug}
-    )
-
-    response = admin_client.get(
-        url,
-        {'region': 'germany', 'service_name': cms.GREAT_INTERNATIONAL}
-    )
-    assert response.status_code == 200
-    assert 'localised_child_pages' in response.json()
-    expected_localised_article = response.json(
-        )['localised_child_pages'][0]['id']
-    assert expected_localised_article == localised_article.pk
-
-
-def test_article_listingpage_without_localised_content(
-        admin_client, international_root_page
-):
-    article_listing_page = factories.InternationalArticleListingPageFactory.create( # NOQA
-        parent=international_root_page,
-        live=True,
-        slug='setup-uk'
-    )
-    factories.InternationalArticlePageFactory.create(
-        parent=article_listing_page,
-        live=True
-    )
-    region_page = factories.InternationalRegionPageFactory(
-        parent=international_root_page,
-        live=True,
-        slug='germany'
-    )
-    regional_folder_page = factories.InternationalLocalisedFolderPageFactory(
-        parent=region_page,
-        live=True,
-        slug='news'  # different branch from setup uk
-    )
-    factories.InternationalArticlePageFactory.create(
-        parent=regional_folder_page,
-        live=True
-    )
-
-    url = reverse(
-        'api:lookup-by-slug',
-        kwargs={'slug': article_listing_page.slug}
-    )
-
-    response = admin_client.get(
-        url,
-        {'region': 'bar', 'service_name': cms.GREAT_INTERNATIONAL}
-    )
-    assert response.status_code == 200
-    assert 'localised_child_pages' in response.json()
-    assert response.json()['localised_child_pages'] == []
-
-
-def test_client_not_passing_region(admin_client, international_root_page):
-    article_listing_page = factories.InternationalArticleListingPageFactory.create(  # NOQA
-        parent=international_root_page,
-        live=True
-    )
-
-    url = reverse(
-        'api:api:pages:detail',
-        kwargs={'pk': article_listing_page.pk}
-    )
-    response = admin_client.get(url)
-    assert response.status_code == 200
-    assert 'localised_child_pages' in response.json()
-    assert response.json()['localised_child_pages'] == []
-
-
+@pytest.mark.django_db
 def test_invest_home_page(document, admin_client, international_root_page):
     page = factories.InvestInternationalHomePageFactory(live=True, parent=international_root_page)
 
@@ -240,6 +148,7 @@ def test_invest_home_page(document, admin_client, international_root_page):
     assert high_potential_ops[0]['title'] == 'Featured'
 
 
+@pytest.mark.django_db
 def test_invest_region_page(admin_client, international_root_page):
     page = factories.InvestRegionPageFactory(
         live=True, featured=True, parent=international_root_page
@@ -252,6 +161,7 @@ def test_invest_region_page(admin_client, international_root_page):
     assert response.status_code == 200
 
 
+@pytest.mark.django_db
 def test_invest_region_landing_page(admin_client, international_root_page):
     page = factories.InvestRegionLandingPageFactory(
         live=True, parent=international_root_page
@@ -265,6 +175,7 @@ def test_invest_region_landing_page(admin_client, international_root_page):
     assert len(response.json()['regions']) == 1
 
 
+@pytest.mark.django_db
 def test_high_potential_opportunity_api(document, admin_client, international_root_page):
     factories.InvestHighPotentialOpportunityDetailPageFactory(
         parent=international_root_page,
@@ -291,9 +202,9 @@ def test_high_potential_opportunity_api(document, admin_client, international_ro
     assert 'other_opportunities' not in parsed['other_opportunities'][0]
 
 
+@pytest.mark.django_db
 def test_international_trade_home_page_exposes_industries(
-        admin_client,
-        international_root_page
+    admin_client, international_root_page
 ):
     industry = factories.InternationalSectorPageFactory(parent=international_root_page,
                                                         live=True)
@@ -309,6 +220,7 @@ def test_international_trade_home_page_exposes_industries(
     assert response.json()['industries'][0]['meta']['pk'] == industry.pk
 
 
+@pytest.mark.django_db
 def test_sector_page_exposes_articles_child_sub_pages(admin_client, international_root_page):
     sector_page = factories.InternationalSectorPageFactory(parent=international_root_page, slug='sector-one')
     factories.InternationalArticlePageFactory(parent=sector_page)