+++ CHANGELOG.md
@@ -10,9 +10,15 @@
 - CI-475 - Added featured description to About DIT services page to be used on featured cards
 - CI-476 - Allow invest regions pages to be added as related regions to About UK pages
 - CI-477 - Why choose the UK page can add international article page as child
+- XOT-1033 - Great domestic homepage
 - CI-478 - Added breadcrumbs label to capital invest landing page
 - CI-474 - Allowed pages to be moved to reflect new IA, changed Invest and About UK landing pages children pages
+- CI-451 - Added new fields to International home page
+- CI-479 - Added video to Why choose UK Page
+- CI-480 - Added SEOOptimisationPanel to more pages
 
+- No ticket - Improve speed of test execution
+- 
 ## [2019.08.27_1](https://github.com/uktrade/directory-cms/releases/tag/2019.08.27_1)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.08.27...2019.08.27_1)
 
@@ -26,6 +32,9 @@
 - CMS-1832 - Update django
 - CMS-1753 - Remove slug links from markdown fields
 - CI-431 - Added `AboutUkRegionListingPage` and `AboutUkRegionPage` which will replace `CapitalInvestRegionPage`
+- CI-400 - Added related regions and text on About Uk pages for map 
+- XOT-1020 - Added extra fields to article page type (domestic and international)
+- CMS-1716 - Remove obsolete code
 - CI-400 - Added related regions and text on About Uk pages for map
 
 ### Fixed Bugs